[feat] Add weight in AStar algorithm
Put heuristic and total cost to double
Creat weight in SearchAlgo

+++ include/Heuristics.hpp
@@ -6,10 +6,10 @@
 #include <tuple>
 #include <vector>
 
-typedef size_t (*heuristic)(const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
+typedef double (*heuristic)(const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
 
-size_t manhattan_distance( const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
-size_t euclidean_heuristic(const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
-size_t tiles_out_of_place( const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
+double manhattan_distance( const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
+double euclidean_heuristic(const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
+double tiles_out_of_place( const std::vector<int>& puzzleNumbers, size_t sizeLine, const std::vector<int>&vecSolution, const std::vector<size_t>& mapSolution);
 
 #endif
